# app/api/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Тестовое REST API приложение: «Добавление товара в заказ»",
    "summary": "Взаимодействие с пользователем посредством HTTP запросов к API серверус использованием статического API ключа.\n\nВсе ответы в формате JSON.",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v0/client/count": {
      "get": {
        "tags": [
          "Клиенты"
        ],
        "summary": "Список клиентов",
        "description": "Метод возвращает список клиентов",
        "operationId": "get_all_clients_api_v0_client_count_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Количество клиентов в ответе",
              "default": 5,
              "title": "Limit"
            },
            "description": "Количество клиентов в ответе"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Смещение относительно первого клиента",
              "default": 0,
              "title": "Offset"
            },
            "description": "Смещение относительно первого клиента"
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Порядок вывода ответа:\n\n```create_date``` - по времени создания\n\n```id``` - по ID клиента",
              "default": "id",
              "title": "Order"
            },
            "description": "Порядок вывода ответа:\n\n```create_date``` - по времени создания\n\n```id``` - по ID клиента"
          },
          {
            "name": "direction",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Порядок сортировки:\n\n```desc``` - от большего к меньшему\n\n```asc``` - от меньшего к большему",
              "default": "desc",
              "title": "Direction"
            },
            "description": "Порядок сортировки:\n\n```desc``` - от большего к меньшему\n\n```asc``` - от меньшего к большему"
          }
        ],
        "responses": {
          "200": {
            "description": "Успешно",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Client"
                  },
                  "title": "Response Get All Clients Api V0 Client Count Get"
                },
                "example": [
                  {
                    "id": 1,
                    "name": "ИП Компания № 1",
                    "create_date": "2025-09-09T06:35:28",
                    "address": "Москва, ул. Трофимова, 32к1"
                  },
                  {
                    "id": 2,
                    "name": "ООО Компания № 2",
                    "create_date": "2025-09-09T06:35:28",
                    "address": "Москва, ул. Каширское шоссе, 4к3"
                  }
                ]
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Invalid or missing API Key"
                }
              }
            },
            "description": "Не авторизован"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "detail": [
                    {
                      "loc": [
                        "string",
                        0],
                      "msg": "string",
                      "type": "string"
                    }
                  ]
                }
              }
            },
            "description": "Ошибка валидации входных данных"
          }
        }
      }
    },
    "/api/v0/product/count": {
      "get": {
        "tags": [
          "Товары"
        ],
        "summary": "Список товаров",
        "description": "Метод возвращает список товаров",
        "operationId": "get_all_products_api_v0_product_count_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Количество товара в ответе",
              "default": 5,
              "title": "Limit"
            },
            "description": "Количество товара в ответе"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Смещение относительно первого товара",
              "default": 0,
              "title": "Offset"
            },
            "description": "Смещение относительно первого товара"
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Порядок вывода ответа:\n\n```create_date``` - по времени создания\n\n```id``` - по ID товара",
              "default": "id",
              "title": "Order"
            },
            "description": "Порядок вывода ответа:\n\n```create_date``` - по времени создания\n\n```id``` - по ID товара"
          },
          {
            "name": "direction",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Порядок сортировки:\n\n```desc``` - от большего к меньшему\n\n```asc``` - от меньшего к большему",
              "default": "desc",
              "title": "Direction"
            },
            "description": "Порядок сортировки:\n\n```desc``` - от большего к меньшему\n\n```asc``` - от меньшего к большему"
          }
        ],
        "responses": {
          "200": {
            "description": "Успешно",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  },
                  "title": "Response Get All Products Api V0 Product Count Get"
                },
                "example": [
                  {
                    "id": 1,
                    "name": "Холодильник Aceline",
                    "category_id": 4,
                    "stock_quantity": 9,
                    "price": 18999,
                    "create_date": "2025-09-09T06:35:28"
                  },
                  {
                    "id": 2,
                    "name": "Холодильник Indesit",
                    "category_id": 4,
                    "stock_quantity": 4,
                    "price": 26799,
                    "create_date": "2025-09-09T06:35:28"
                  }
                ]
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Invalid or missing API Key"
                }
              }
            },
            "description": "Не авторизован"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "detail": [
                    {
                      "loc": [
                        "string",
                        0],
                      "msg": "string",
                      "type": "string"
                    }
                  ]
                }
              }
            },
            "description": "Ошибка валидации входных данных"
          }
        }
      }
    },
    "/api/v0/order/create/": {
      "post": {
        "tags": [
          "Заказы"
        ],
        "summary": "Создать заказ",
        "description": "Метод создает новый заказ",
        "operationId": "create_order_api_v0_order_create__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrder"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Успешно",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                },
                "example": {
                  "id": 7,
                  "client_id": 3,
                  "create_date": "2025-09-09T08:55:56"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка валидации входных данных",
            "content": {
              "application/json": {
                "example": {
                  "detail": [
                    {
                      "loc": [
                        "string",
                        0],
                      "msg": "string",
                      "type": "string"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Неверный запрос",
            "content": {
              "application/json": {
                "example": {
                  "detail": "The product/s is out of stock or unavailable with: product_id = 25"
                }
              }
            }
          }
        }
      }
    },
    "/api/v0/order/upload/": {
      "post": {
        "tags": [
          "Заказы"
        ],
        "summary": "Добавить товар в заказ",
        "description": "Метод добавляет товар в заказ",
        "operationId": "update_order_api_v0_order_upload__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrder"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Успешно",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                },
                "example": {
                  "id": 7,
                  "client_id": 3
                }
              }
            }
          },
          "422": {
            "description": "Ошибка валидации входных данных",
            "content": {
              "application/json": {
                "example": {
                  "detail": [
                    {
                      "loc": [
                        "string",
                        0],
                      "msg": "string",
                      "type": "string"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Неверный запрос",
            "content": {
              "application/json": {
                "example": {
                  "detail": "The product/s is out of stock or unavailable with: product_id = 25"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Category": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "create_date": {
            "type": "string",
            "format": "date-time",
            "title": "Create Date"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "parent": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Category"
              },
              {
                "type": "null"
              }
            ]
          },
          "subcategories": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Category"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subcategories"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "create_date"
        ],
        "title": "Category"
      },
      "Client": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "create_date": {
            "type": "string",
            "format": "date-time",
            "title": "Create Date"
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "orders": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Order"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Orders"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "create_date"
        ],
        "title": "Client"
      },
      "CreateOrder": {
        "properties": {
          "client_id": {
            "type": "integer",
            "title": "Client Id"
          },
          "products": {
            "items": {
              "$ref": "#/components/schemas/ProductInOrder"
            },
            "type": "array",
            "title": "Products"
          }
        },
        "type": "object",
        "required": [
          "client_id",
          "products"
        ],
        "title": "CreateOrder"
      },
      "Order": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "client_id": {
            "type": "integer",
            "title": "Client Id"
          },
          "create_date": {
            "type": "string",
            "format": "date-time",
            "title": "Create Date"
          },
          "client": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Client"
              },
              {
                "type": "null"
              }
            ]
          },
          "items": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/OrderItem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "id",
          "client_id",
          "create_date"
        ],
        "title": "Order"
      },
      "OrderItem": {
        "properties": {
          "order_id": {
            "type": "integer",
            "title": "Order Id"
          },
          "product_id": {
            "type": "integer",
            "title": "Product Id"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "unit_price": {
            "type": "number",
            "title": "Unit Price"
          },
          "order": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Order"
              },
              {
                "type": "null"
              }
            ]
          },
          "product": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Product"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "order_id",
          "product_id",
          "quantity",
          "unit_price"
        ],
        "title": "OrderItem"
      },
      "Product": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "category_id": {
            "type": "integer",
            "title": "Category Id"
          },
          "stock_quantity": {
            "type": "integer",
            "title": "Stock Quantity"
          },
          "price": {
            "type": "number",
            "title": "Price"
          },
          "create_date": {
            "type": "string",
            "format": "date-time",
            "title": "Create Date"
          },
          "category": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Category"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "category_id",
          "stock_quantity",
          "price",
          "create_date"
        ],
        "title": "Product"
      },
      "ProductInOrder": {
        "properties": {
          "product_id": {
            "type": "integer",
            "title": "Product Id"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          }
        },
        "type": "object",
        "required": [
          "product_id",
          "quantity"
        ],
        "title": "ProductInOrder"
      },
      "UpdateOrder": {
        "properties": {
          "order_id": {
            "type": "integer",
            "title": "Order Id"
          },
          "products": {
            "items": {
              "$ref": "#/components/schemas/ProductInOrder"
            },
            "type": "array",
            "title": "Products"
          }
        },
        "type": "object",
        "required": [
          "order_id",
          "products"
        ],
        "title": "UpdateOrder"
      }
    }
  }
}